added user PostScore
I am working on the userQuery_click method that I created. I am having an issue where I am trying to do the same thing as what i did in the last method. I have the skeleton for the low score. If you can see what is the issue that will be great.